+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Parse server-side validation errors.
 - Add support for readonly attributes in serialization
 - Basic serialization support via `as_json` and `serializable_hash` helpers